Convert WearableListenerServices to new filterable approach in Google Play services 8.3
Rather than receiving all data/messages in a single WearableListenerServices, Google Play services 8.3 allows filtering to only specific messages/data APIs, simplifying our logic.
