+++ main/src/main/java/com/google/android/apps/muzei/wearable/NotificationOpenListenerService.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2014 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.apps.muzei.wearable;
+
+import android.content.Intent;
+import android.content.pm.PackageManager;
+
+import com.google.android.gms.wearable.MessageEvent;
+import com.google.android.gms.wearable.WearableListenerService;
+
+public class NotificationOpenListenerService extends WearableListenerService {
+    @Override
+    public void onMessageReceived(MessageEvent messageEvent) {
+        // Only notification/open actions trigger this WearableListenerService
+        PackageManager packageManager = getPackageManager();
+        Intent mainIntent = packageManager.getLaunchIntentForPackage(getPackageName());
+        startActivity(mainIntent);
+    }
+}
