--- wearable/src/main/java/com/google/android/apps/muzei/MuzeiWearableListenerService.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2014 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.android.apps.muzei;
-
-import android.content.Intent;
-
-import com.google.android.gms.wearable.DataEvent;
-import com.google.android.gms.wearable.DataEventBuffer;
-import com.google.android.gms.wearable.WearableListenerService;
-
-/**
- * WearableListenerService responsible to receiving Data Layer changes with updated artwork
- */
-public class MuzeiWearableListenerService extends WearableListenerService {
-    @Override
-    public void onDataChanged(final DataEventBuffer dataEvents) {
-        for (DataEvent dataEvent : dataEvents) {
-            if (dataEvent.getType() != DataEvent.TYPE_CHANGED) {
-                continue;
-            }
-            if (dataEvent.getDataItem().getUri().getPath().equals("/artwork")) {
-                // Kick off a service responsible for reloading the current artwork
-                startService(new Intent(this, ArtworkCacheIntentService.class));
-            }
-        }
-    }
-}
